refactor into realize
